basic implementation 😼😼😼😼 (supporting saving and clicks)
This first version of our extension includes the following features:
    1. A small pop-up with a game.
    2. Local saving for the game with chrome.storage.local.
The game is a simple clicker game which I stylized with the CSS found inside main.css.
The code for the game and the saving feature can be found within click.js.
The pop-up can be found in popup.html.
It's important to note that all of these are found within the main folder that contains everything.

Every click results in a call to chrome.storage to save. This didn't seem to be an issue within testing. Later, we can debounce if needed, or make use of an on close event.